NAMESPACE VALIDATOR ADDED

+++ pkg/storage/storage.go
@@ -1,6 +1,7 @@
 package storage
 
 import (
+	"github.com/golang/protobuf/proto"
 	"github.com/mkawserm/flamed/pkg/iface"
 	"github.com/mkawserm/flamed/pkg/pb"
 	"github.com/mkawserm/flamed/pkg/utility"
@@ -93,6 +94,10 @@ func (s *Storage) IsExists(namespace []byte, key []byte) bool {
 }
 
 func (s *Storage) Read(namespace []byte, key []byte) ([]byte, error) {
+	if !utility.IsNamespaceValid(namespace) {
+		return nil, x.ErrInvalidNamespace
+	}
+
 	d, err := s.mKVStorage.Read(namespace, key)
 	if err == x.ErrUidDoesNotExists {
 		return nil, nil
@@ -101,26 +106,52 @@ func (s *Storage) Read(namespace []byte, key []byte) ([]byte, error) {
 }
 
 func (s *Storage) Delete(namespace []byte, key []byte) error {
+	if !utility.IsNamespaceValid(namespace) {
+		return x.ErrInvalidNamespace
+	}
+
 	return s.mKVStorage.Delete(namespace, key)
 }
 
 func (s *Storage) Create(namespace []byte, key []byte, value []byte) error {
+	if !utility.IsNamespaceValid(namespace) {
+		return x.ErrInvalidNamespace
+	}
+
 	return s.mKVStorage.Create(namespace, key, value)
 }
 
 func (s *Storage) Update(namespace []byte, key []byte, value []byte) error {
+	if !utility.IsNamespaceValid(namespace) {
+		return x.ErrInvalidNamespace
+	}
+
 	return s.mKVStorage.Update(namespace, key, value)
 }
 
 func (s *Storage) Append(namespace []byte, key []byte, value []byte) error {
+	if !utility.IsNamespaceValid(namespace) {
+		return x.ErrInvalidNamespace
+	}
+
 	return s.mKVStorage.Append(namespace, key, value)
 }
 
 func (s *Storage) ApplyBatchAction(batch *pb.FlameBatchAction) error {
+	for idx := range batch.FlameActionList {
+		if !utility.IsNamespaceValid(batch.FlameActionList[idx].FlameEntry.Namespace) {
+			return x.ErrInvalidNamespace
+		}
+	}
+
 	return s.mKVStorage.ApplyBatchAction(batch)
 }
 
 func (s *Storage) ApplyAction(action *pb.FlameAction) error {
+	if !utility.IsNamespaceValid(action.FlameEntry.Namespace) {
+		return x.ErrInvalidNamespace
+	}
+
 	return s.mKVStorage.ApplyAction(action)
 }
 
@@ -145,10 +176,18 @@ func (s *Storage) QueryAppliedIndex() (uint64, error) {
 }
 
 func (s *Storage) AddIndexMeta(meta *pb.FlameIndexMeta) error {
+	if !utility.IsNamespaceValid(meta.Namespace) {
+		return x.ErrInvalidNamespace
+	}
+
 	return s.mKVStorage.AddIndexMeta(meta)
 }
 
 func (s *Storage) GetIndexMeta(meta *pb.FlameIndexMeta) error {
+	if !utility.IsNamespaceValid(meta.Namespace) {
+		return x.ErrInvalidNamespace
+	}
+
 	return s.mKVStorage.GetIndexMeta(meta)
 }
 
@@ -157,10 +196,18 @@ func (s *Storage) GetAllIndexMeta() ([]*pb.FlameIndexMeta, error) {
 }
 
 func (s *Storage) UpdateIndexMeta(meta *pb.FlameIndexMeta) error {
+	if !utility.IsNamespaceValid(meta.Namespace) {
+		return x.ErrInvalidNamespace
+	}
+
 	return s.mKVStorage.UpdateIndexMeta(meta)
 }
 
 func (s *Storage) DeleteIndexMeta(meta *pb.FlameIndexMeta) error {
+	if !utility.IsNamespaceValid(meta.Namespace) {
+		return x.ErrInvalidNamespace
+	}
+
 	return s.mKVStorage.DeleteIndexMeta(meta)
 }
 
@@ -185,10 +232,16 @@ func (s *Storage) DeleteUser(user *pb.FlameUser) error {
 }
 
 func (s *Storage) AddAccessControl(ac *pb.FlameAccessControl) error {
+	if !utility.IsNamespaceValid(ac.Namespace) {
+		return x.ErrInvalidNamespace
+	}
 	return s.mKVStorage.AddAccessControl(ac)
 }
 
 func (s *Storage) GetAccessControl(ac *pb.FlameAccessControl) error {
+	if !utility.IsNamespaceValid(ac.Namespace) {
+		return x.ErrInvalidNamespace
+	}
 	return s.mKVStorage.GetAccessControl(ac)
 }
 
@@ -197,13 +250,70 @@ func (s *Storage) GetAllAccessControl() ([]*pb.FlameAccessControl, error) {
 }
 
 func (s *Storage) UpdateAccessControl(ac *pb.FlameAccessControl) error {
+	if !utility.IsNamespaceValid(ac.Namespace) {
+		return x.ErrInvalidNamespace
+	}
 	return s.mKVStorage.UpdateAccessControl(ac)
 }
 
 func (s *Storage) DeleteAccessControl(ac *pb.FlameAccessControl) error {
+	if !utility.IsNamespaceValid(ac.Namespace) {
+		return x.ErrInvalidNamespace
+	}
 	return s.mKVStorage.DeleteAccessControl(ac)
 }
 
 func (s *Storage) ApplyProposal(pp *pb.FlameProposal) error {
-	return s.mKVStorage.ApplyProposal(pp)
+	if pp.FlameProposalType == pb.FlameProposal_BATCH_ACTION {
+		batchAction := &pb.FlameBatchAction{}
+		if err := proto.Unmarshal(pp.FlameProposalData, batchAction); err != nil {
+			return x.ErrFailedToApplyProposal
+		}
+
+		return s.ApplyBatchAction(batchAction)
+	} else if pp.FlameProposalType == pb.FlameProposal_ADD_INDEX_META {
+		indexMeta := &pb.FlameIndexMeta{}
+		if err := proto.Unmarshal(pp.FlameProposalData, indexMeta); err != nil {
+			return x.ErrFailedToApplyProposal
+		}
+
+		return s.AddIndexMeta(indexMeta)
+	} else if pp.FlameProposalType == pb.FlameProposal_UPDATE_INDEX_META {
+		indexMeta := &pb.FlameIndexMeta{}
+		if err := proto.Unmarshal(pp.FlameProposalData, indexMeta); err != nil {
+			return x.ErrFailedToApplyProposal
+		}
+
+		return s.UpdateIndexMeta(indexMeta)
+	} else if pp.FlameProposalType == pb.FlameProposal_DELETE_INDEX_META {
+		indexMeta := &pb.FlameIndexMeta{}
+		if err := proto.Unmarshal(pp.FlameProposalData, indexMeta); err != nil {
+			return x.ErrFailedToApplyProposal
+		}
+
+		return s.DeleteIndexMeta(indexMeta)
+	} else if pp.FlameProposalType == pb.FlameProposal_ADD_ACCESS_CONTROL {
+		ac := &pb.FlameAccessControl{}
+		if err := proto.Unmarshal(pp.FlameProposalData, ac); err != nil {
+			return x.ErrFailedToApplyProposal
+		}
+
+		return s.AddAccessControl(ac)
+	} else if pp.FlameProposalType == pb.FlameProposal_UPDATE_ACCESS_CONTROL {
+		ac := &pb.FlameAccessControl{}
+		if err := proto.Unmarshal(pp.FlameProposalData, ac); err != nil {
+			return x.ErrFailedToApplyProposal
+		}
+
+		return s.UpdateAccessControl(ac)
+	} else if pp.FlameProposalType == pb.FlameProposal_DELETE_ACCESS_CONTROL {
+		ac := &pb.FlameAccessControl{}
+		if err := proto.Unmarshal(pp.FlameProposalData, ac); err != nil {
+			return x.ErrFailedToApplyProposal
+		}
+
+		return s.DeleteAccessControl(ac)
+	}
+
+	return x.ErrFailedToApplyProposal
 }
